SIANXSVC-1123: Refactor library to use generics
This change introduces a lot of API changes. First of all, we got rid of
the `reflect`-based approach of the v1. The primary reason for this was
compile-time safety of the handlers. Although a first approach used a
`http.Handler`-like interface for that, this was converted to a
polymorphic (commonly known as generics) solution, with the pattern
inspired by the open-source River library [1].

The basic idea here is to wrap the typed handler in a non-typed struct.

[1]: https://github.com/riverqueue/river

BREAKING: Increases the minimum version to Go 1.18 and is incompatible
          with the old API.

Co-authored-by: Patrick Taibel <ptaibel@anexia-it.com>